Output NetworkInfo from Dynamic Honey Badger
This allows the user to use current cryptographic keys and key shares.

An example use case would be if we want to make the sequence of batches
verifiable for third parties, like a blockchain. We would have each
batch signed by the set of validators that produced it, and whenever a
validator changes, sign the new public key using the old key shares.

To a third party that would then look like the owner of the secret key
signed each batch, and occasionally changed their key.