Add `.rustfmt.toml` and improve formatting.
This gets us much closer to what is fairly commonplace Rust
formatting. I'm not sure why the rustfmt defaults are set the way
they are, this is a huge improvement. I still really dislike how it
indents vertical lists of arguments so far to the right but disabling
that option makes the bracket style far to ugly and hard to read. One
day maybe I'll look into the rustfmt source and see if I can add an
option to reduce maximum indents to +2 or something.

If we decide to use this configuration (which we will because we are
sane and intelligent individuals), authors of PRs can just copy .rustfmt.toml
into their branches and run cargo fmt (to prevent death by merge conflict).